fixing bugs

+++ Libft/ft_strjoin.c
@@ -6,7 +6,7 @@
 /*   By: eel-ghan <eel-ghan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/06 10:43:34 by eel-ghan          #+#    #+#             */
-/*   Updated: 2021/11/14 12:44:40 by eel-ghan         ###   ########.fr       */
+/*   Updated: 2021/11/17 13:06:56 by eel-ghan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,9 +31,9 @@ static char	*ft_strcat(char *dest, char *src)
 
 char	*ft_strjoin(char const *s1, char const *s2)
 {
-	char			*str;
-	unsigned int	s1_len;
-	unsigned int	s2_len;
+	char		*str;
+	size_t		s1_len;
+	size_t		s2_len;
 
 	if (!s1 || !s2)
 		return (0);
